pdf_execl_word/requirements.txt
# pdf2word
pdfminer3k
python-docx
# word2pdf
comtypes